+++ lib/publishing.ts
@@ -358,6 +358,8 @@ export class PublishToGitHub extends cdk.Construct implements IPublisher {
     const forReal = props.dryRun === undefined ? 'false' : (!props.dryRun).toString();
     this.additionalInputArtifacts = props.additionalInputArtifacts;
 
+    const githubToken = cdk.SecretValue.secretsManager(props.githubRepo.tokenSecretArn);
+
     const shellable = new Shellable(this, 'Default', {
       platform: new LinuxPlatform(cbuild.LinuxBuildImage.UBUNTU_14_04_NODEJS_10_1_0),
       scriptDirectory: path.join(__dirname, 'publishing', 'github'),
@@ -366,7 +368,7 @@ export class PublishToGitHub extends cdk.Construct implements IPublisher {
         BUILD_MANIFEST: props.buildManifestFileName || './build.json',
         CHANGELOG: props.changelogFileName || './CHANGELOG.md',
         SIGNING_KEY_ARN: props.signingKey.credential.secretArn,
-        GITHUB_TOKEN: props.githubRepo.token.toString(),
+        GITHUB_TOKEN: githubToken.toString(),
         GITHUB_OWNER: props.githubRepo.owner,
         GITHUB_REPO: props.githubRepo.repo,
         FOR_REAL: forReal,