+++ lib/repo.ts
@@ -3,23 +3,24 @@ import { core as core, aws_codebuild as cbuild, aws_codecommit as ccommit,
   core as cdk } from "monocdk-experiment";
 import { ExternalSecret } from "./permissions";
 
-
-
-
-
 export interface IRepo {
   repositoryUrlHttp: string;
   repositoryUrlSsh: string;
   readonly allowsBadge: boolean;
-  readonly token: core.SecretValue | undefined;
-  createBuildSource(parent: cdk.Construct, webhook: boolean, branch?: string): cbuild.ISource;
+  readonly tokenSecretArn?: string;
+  createBuildSource(parent: cdk.Construct, webhook: boolean, options?: BuildSourceOptions): cbuild.ISource;
   createSourceStage(pipeline: cpipeline.Pipeline, branch: string): cpipeline.Artifact;
   describe(): any;
 }
 
+export interface BuildSourceOptions {
+  branch?: string;
+  cloneDepth?: number;
+}
+
 export class CodeCommitRepo implements IRepo {
   public readonly allowsBadge = false;
-  public readonly token: core.SecretValue | undefined = undefined;
+  public readonly tokenSecretArn?: string;
 
   constructor(private readonly repository: ccommit.IRepository) {
 
@@ -47,9 +48,10 @@ export class CodeCommitRepo implements IRepo {
     return this.repository.repositoryCloneUrlSsh;
   }
 
-  public createBuildSource(_: cdk.Construct, _webhook: boolean): cbuild.ISource {
+  public createBuildSource(_: cdk.Construct, _webhook: boolean, options: BuildSourceOptions = { }): cbuild.ISource {
     return cbuild.Source.codeCommit({
       repository: this.repository,
+      cloneDepth: options.cloneDepth,
     });
   }
 
@@ -62,7 +64,7 @@ interface GitHubRepoProps {
   /**
    * The OAuth token secret that allows access to your github repo.
    */
-  token: cdk.SecretValue;
+  tokenSecretArn: string;
 
   /**
    * In the form "account/repo".
@@ -74,7 +76,7 @@ export class GitHubRepo implements IRepo {
   public readonly allowsBadge = true;
   public readonly owner: string;
   public readonly repo: string;
-  public readonly token: cdk.SecretValue;
+  public readonly tokenSecretArn: string;
 
   constructor(props: GitHubRepoProps) {
     const repository = props.repository;
@@ -82,8 +84,7 @@ export class GitHubRepo implements IRepo {
 
     this.owner = owner;
     this.repo = repo;
-
-    this.token = props.token;
+    this.tokenSecretArn = props.tokenSecretArn;
   }
 
   public get repositoryUrlHttp() {
@@ -101,7 +102,7 @@ export class GitHubRepo implements IRepo {
     stage.addAction(new cpipeline_actions.GitHubSourceAction({
       actionName: 'Pull',
       branch,
-      oauthToken: this.token,
+      oauthToken: cdk.SecretValue.secretsManager(this.tokenSecretArn),
       owner: this.owner,
       repo: this.repo,
       output: sourceOutput,
@@ -109,17 +110,17 @@ export class GitHubRepo implements IRepo {
     return sourceOutput;
   }
 
-  public createBuildSource(_: cdk.Construct, webhook: boolean, branch?: string): cbuild.ISource {
+  public createBuildSource(_: cdk.Construct, webhook: boolean, options: BuildSourceOptions = { }): cbuild.ISource {
     return cbuild.Source.gitHub({
       owner: this.owner,
       repo: this.repo,
       webhook,
+      cloneDepth: options.cloneDepth,
       reportBuildStatus: webhook,
       webhookFilters: webhook
-          ? this.createWebhookFilters(branch)
+          ? this.createWebhookFilters(options.branch)
           : undefined,
     });
-
   }
 
   public describe() {