+++ pipeline/delivlib.ts
@@ -16,10 +16,10 @@ export class DelivLibPipelineStack extends cdk.Stack {
 
     const github = new delivlib.WritableGitHubRepo({
       repository: 'awslabs/aws-delivlib',
-      token: cdk.SecretValue.secretsManager('github-token'),
+      tokenSecretArn: 'arn:aws:secretsmanager:us-east-1:712950704752:secret:github-token-nnAqfW',
       commitEmail: 'aws-cdk-dev+delivlib@amazon.com',
       commitUsername: 'aws-cdk-dev',
-      sshKeySecret: { secretArn: 'arn:aws:secretsmanager:us-east-1:712950704752:secret:delivlib/github-ssh-lwzfjW' }
+      sshKeySecret: { secretArn: 'arn:aws:secretsmanager:us-east-1:712950704752:secret:awslabs/delivlib/github-ssh-UBHEyF' }
     });
 
     const pipeline = new delivlib.Pipeline(this, 'GitHubPipeline', {